+++ src/components/Chats.jsx
@@ -1,8 +1,12 @@
-import React, { useContext, useState } from 'react'
+import React, { useContext } from 'react'
 import { Box, IconButton, Stack, Typography, InputBase, styled, Button, Divider, Avatar, Badge } from '@mui/material'
 import { ArchiveBox, CircleDashed, MagnifyingGlass } from 'phosphor-react'
 import { faker } from '@faker-js/faker';
 import { AppContext } from '../Context/ParentContext';
+import Search from './Search/Search';
+import SearchIconWrapper from './Search/SearchIconWrapper';
+import StyledInputBase from './Search/StyledInputBase';
+import ChatElement from './Chats/ChatElement';
 
 const ChatList = [
     {
@@ -87,95 +91,6 @@ const ChatList = [
     },
 ];
 
-const StyledBadge = styled(Badge)(({ theme }) => ({
-    '& .MuiBadge-badge': {
-        backgroundColor: '#44b700',
-        color: '#44b700',
-        boxShadow: `0 0 0 2px ${theme.palette.background.paper}`,
-        '&::after': {
-            position: 'absolute',
-            top: 0,
-            left: 0,
-            width: '100%',
-            height: '100%',
-            borderRadius: '50%',
-            animation: 'ripple 1.2s infinite ease-in-out',
-            border: '1px solid currentColor',
-            content: '""',
-        },
-    },
-    '@keyframes ripple': {
-        '0%': {
-            transform: 'scale(.8)',
-            opacity: 1,
-        },
-        '100%': {
-            transform: 'scale(2.4)',
-            opacity: 0,
-        },
-    },
-}));
-
-const ChatElement = ({ id, name, img, msg, time, unread, online }) => {
-
-    const {isToggled} = useContext(AppContext)
-
-    return (
-        <Box p={2} sx={{ width: "90%", borderRadius: 1, backgroundColor: isToggled ? "#fff" : "#1F2631" }}>
-            <Stack direction="row" alignItems={"center"} justifyContent="space-between">
-                <Stack direction="row" spacing={2}>
-                    {online ? <StyledBadge overlap="circular" anchorOrigin={{ vertical: "bottom", horizontal: "right" }} variant="dot" >
-                        <Avatar src={img} />
-                    </StyledBadge> : <Avatar src={img} />}
-
-                    <Stack spacing={0.3}>
-                        <Typography sx={{ color: isToggled ?"#000" : "#fff" }} variant='subtitle2'>
-                            {name}
-                        </Typography>
-                        <Typography sx={{ color: isToggled ?"#000" : "#fff" }} variant='caption'>
-                            {msg}
-                        </Typography>
-                    </Stack>
-                </Stack>
-                <Stack spacing={2} alignItems={"center"}>
-                    <Typography sx={{ fontWeight: 600, color: isToggled ?"#000" : "#fff" }} variant='caption' >
-                        {time}
-                    </Typography>
-                    <Badge color="primary" badgeContent={unread}>
-                    </Badge>
-                </Stack>
-            </Stack>
-        </Box>
-    )
-}
-
-const Search = styled("div")(({ theme }) => ({
-    position: "relative",
-    marginRight: theme.spacing(2),
-    marginLeft: 0,
-    width: "100%",
-}))
-
-const SearchIconWrapper = styled("div")(({ theme }) => ({
-    padding: theme.spacing(0, 2),
-    height: "100%",
-    position: "absolute",
-    pointerEvents: "none",
-    zIndex: 1000,
-    display: "flex",
-    alignItems: "center",
-    justifyContent: "center",
-}))
-
-const StyledInputBase = styled(InputBase)(({ theme }) => ({
-    "& .MuiInputBase-input": {
-        padding: theme.spacing(1, 1, 1, 0),
-        // vertical padding + font size from search Icon
-        paddingLeft: `calc(1em + ${theme.spacing(4)})`,
-        width: "100%",
-    },
-}))
-
 function Chats() {
 
     const {isToggled} = useContext(AppContext)
@@ -196,7 +111,7 @@ function Chats() {
                         <SearchIconWrapper>
                             <MagnifyingGlass color={isToggled ? "#709CE6" : "grey"} />
                         </SearchIconWrapper>
-                        <StyledInputBase sx={{ color: isToggled ? "#36454F" : "	#fff", backgroundColor: isToggled ? "#fff" : "#161A21", borderRadius: 20 }} placeholder="Search..." inputProps={{ "aria-label": "search" }} />
+                        <StyledInputBase sx={{ color: isToggled ? "#36454F" : "	#fff", backgroundColor: isToggled ? "#fff" : "#1F2631", borderRadius: 20 }} placeholder="Search..." inputProps={{ "aria-label": "search" }} />
                     </Search>
                 </Stack>
                 <Stack spacing={1}>
@@ -228,6 +143,9 @@ function Chats() {
                 </Stack>
             </Stack>
         </Box>
+
+        // Right
+        // TODO => Reuse Conversation Components
     )
 }
