+++ src/components/Chats/index.jsx
@@ -1,0 +1,2 @@
+import {default as StyledBadge} from './StyledBadge'
+import {default as ChatElement} from './ChatElement'
