+++ src/components/Search/Index.jsx
@@ -1,0 +1,3 @@
+import { default as Search } from './Search'
+import { default as SearchIconWrapper } from './SearchIconWrapper'
+import { default as StyledInputBase } from './StyledInputBase'
