+++ src/components/Search/Search.jsx
@@ -1,0 +1,10 @@
+import { styled } from "@mui/material"
+
+const Search = styled("div")(({ theme }) => ({
+    position: "relative",
+    marginRight: theme.spacing(2),
+    marginLeft: 0,
+    width: "100%",
+}))
+
+export default Search
